fix(parser): handle the error in building the 'fcn_form' syntactical structure
To build the 'fcn_form' syntactical structure, after accepting the '<IDENTIFIER>', more than one function shall be called.
Thus the queue shall be dequeued after adding '<IDENTIFIER>' as the left child of the 'fcn_form'.
But it is not dequeued.
Fix this issue by dequeing the queue after adding '<IDENTIFIER>' as the left child of the 'fcn_form'.